move split commands test to `test_bash_parsing`
